ML.NeuralNetwork

Documentation

data NeuralNetwork

Constructors

NeuralNetwork

Fields

config :: [Int]

weights :: [Matrix Float]

biases :: [Matrix Float]

createNeuralNetwork

Arguments

:: [Int]
List of Nodes (Int) per Layer

-> NeuralNetwork

The function createNeuralNetwork creates a randomly initialised network with the specified layers

predict :: NeuralNetwork -> Matrix Float -> Matrix Float

predict takes an network and a fitting input and runs a forwardPass with these parameters. The resulting output matrix is returned

forwardPass :: NeuralNetwork -> Matrix Float -> [Matrix Float]

forwardPass' :: [Matrix Float] -> [Matrix Float] -> Matrix Float -> [Matrix Float]

train :: NeuralNetwork -> [(Matrix Float, Matrix Float)] -> Float -> IO NeuralNetwork

train' :: NeuralNetwork -> [(Matrix Float, Matrix Float)] -> Float -> Float -> Int -> IO NeuralNetwork

data BackpropR

Constructors

BackpropR

Fields

nn :: NeuralNetwork

totalError :: Float

totalIterations :: Int

backprop :: NeuralNetwork -> Matrix Float -> Matrix Float -> Float -> Float -> Int -> IO BackpropR

update :: NeuralNetwork -> [(Matrix Float, Matrix Float)] -> Float -> NeuralNetwork

update' :: [Matrix Float] -> [Matrix Float] -> Float -> [Matrix Float]

gradients :: [Matrix Float] -> [Matrix Float] -> [Matrix Float] -> Matrix Float -> [(Matrix Float, Matrix Float)]

serialize :: NeuralNetwork -> FilePath -> IO ()

deserialize :: FilePath -> IO NeuralNetwork

serialize2 :: NeuralNetwork -> FilePath -> IO ()

deserialize2 :: FilePath -> IO NeuralNetwork

shuffleSamples :: [(Matrix Float, Matrix Float)] -> Int -> [(Matrix Float, Matrix Float)]

argmax :: Matrix Float -> Int

toCategorical :: Int -> Int -> [Float]

test :: NeuralNetwork -> [(Matrix Float, Matrix Float)] -> IO ()

randomRMatrix :: Int -> Int -> (Float, Float) -> Matrix Float

zeroMatrix :: Int -> Int -> Matrix Float

multiplyElementwise :: Matrix Float -> Matrix Float -> Matrix Float

sigmoid :: Float -> Float

sigmoid' :: Float -> Float