MNIST

Description

Provides functions for parsing MNIST data, aswell as additional functions for working with PNGSs.

Parsing

getTrainingSamples

Arguments

:: FilePath
path to "train-images.idx3-ubyte"

-> FilePath
path to "train-labels.idx3-ubyte"

-> IO [(Matrix Float, Matrix Float)]
Training data as a list of pairs, where fst represents an image and snd the corresponding label

getTrainingSamples is used to parse the raw MNIST training data to a representation usable in Haskell

getTestSamples

Arguments

:: FilePath
path to "t10k-images.idx3-ubyte"

-> FilePath
path to "t10k-labels.idx3-ubyte"

-> IO [(Matrix Float, Matrix Float)]
Test data as a list of pairs, where fst represents an image and snd the corresponding label

getTestSamples is used to parse the raw MNIST test data to a representation usable in Haskell

pngToVector

Arguments

:: FilePath
path to a .png file

-> IO (Matrix Float)
float matrix representing the input file

pngToVector takes a file path and parses it to an equivalent float matrix

Rendering

vectorToPNG

Arguments

:: Matrix Float
float matrix to write

-> FilePath
path to write the .png to

-> IO ()

vectorToPng takes a float matrix and a file path, creates an image representation of the input matrix