NeuralNetwork

Description

A naive neural-network implementation in haskell.

Use initNN to get an initialised neural network and train it using the train function.

You can get predictions of your trained neural network by running predict.

DataType

data NeuralNetwork

The data type NeuralNetwork represents the state of a network.

It contains:

config (nodes per layer) as a list of Int

weights as Matrix of Float

biases as Matrix of Float

Initialisiation

initNN

Arguments

:: [Int]
List of Nodes per Layer

-> Int
Seed for the random generation of nodes

-> IO NeuralNetwork

createNeuralNetwork creates a randomly initialised network with the specified layers

For example:

A network with 784 input-nodes, 2 hidden layers with 1000 nodes each and 10 output-nodes

> network <- createNeuralNetwork [784,1000,1000,10]

Prediction

predict

Arguments

:: NeuralNetwork
Trained NeuralNetwork that will be used to predict an output

-> Matrix Float
Matrix of the input values for the given network

-> Matrix Float
Matrix of the output node values

predict takes an network and a fitting input and runs a forwardPass with these parameters. The resulting output Matrix is returned

Training

train

Arguments

:: NeuralNetwork
Network

-> [(Matrix Float, Matrix Float)]
List of Tupels of the corresponding in-/output matrices

-> Float
Learing rate

-> IO NeuralNetwork

The function train should be used to train the network with the passed inputs/outputs and the specified learning rate. Returns the updated NeuralNetwork as IO

Serializiation

Binary coded

serialize

Arguments

:: NeuralNetwork
Network that should be saved

-> FilePath
Relative path to the file

-> IO ()

The serialize function is used to write a network to a specified file path as a binary file. Use deserialize to read this file. Alternatively you can use serializePlain and deserializePlain to save/read as text file.

deserialize

Arguments

:: FilePath
Relative path to the file

-> IO NeuralNetwork

Used to load a network from a binary encoded file

Example usage:

> network <- deserialize "networks/network_3.txt"

Plain coded

serializePlain

Arguments

:: NeuralNetwork
Network that should be saved

-> FilePath
Relative path to the file

-> IO ()

Used the same as serialize but saves the network as text instead of binary. Use deserializePlain to read these files!

deserializePlain

Arguments

:: FilePath
Relative path to the file

-> IO NeuralNetwork

Used the same as deserialize but reads the network as text instead of binary.

Helper functions

sigmoid

Arguments

:: Float
x

-> Float
y

Sigmoid function

sigmoid'

Arguments

:: Float
x

-> Float
y

Derivate of Sigmoid funtion

randomRMatrix

Arguments

:: Int
(m) Number of rows

-> Int
(n) Number of columns

-> (Float, Float)
Range of the random numbers e.g. (-1,1)

-> Int
Seed for the random generator

-> IO (Matrix Float)

Generates a IO matrix with the specified dimensions, initialised with a random number in the given range.

zeroMatrix

Arguments

:: Int
(m) Number of rows

-> Int
(n) Number of columns

-> Matrix Float

Generates a matrix with the specified dimensions, initialised 0's.

multiplyElementwise

Arguments

:: Matrix Float
Matrix A

-> Matrix Float
Matrix B

-> Matrix Float

The function multiplyElementwise is used to multiply each element of matrix A with the element at the same position in matrix B.