Util

Description

Util functions for Machine Learning.

Documentation

argmax

Arguments

:: Matrix Float
matrix

-> Int
argmax

argmax takes a Nx1 matrix and returns the index of the highest value | important: this function is undefined for NxM matrices For example:

@> argmax (fromList 4 1 [0,42,1,17]) -- returns 1

toCategorical

Arguments

:: Int
class

-> Int
number of classes

-> Matrix Float
binary class matrix

toCategorical converts a class to a binary class matrix For example:

@> toCategorical 1 3 -- returns [0,1,0]

shuffle

Arguments

:: [a]
unshuffled list

-> Int
seed

-> [a]
shuffled list

shuffle shuffles a list For example:

@> shuffle [1,2,3] 42 -> returns [3,2,1]

reshape

Arguments

:: Matrix a
matrix

-> Int
rows

-> Int

-> Matrix a