NeuralNetwork

Description

A naive haskell neural-network implementation

DataType

data NeuralNetwork

The data type NeuralNetwork represents the state of a network.

It contains:

config (nodes per layer) as a list of Int

weights as Matrix of Float

biases as Matrix of Float

Initialisiation

createNeuralNetwork

Arguments

:: [Int]
List of Nodes (Int) per Layer

-> NeuralNetwork

The function createNeuralNetwork creates a randomly initialised network with the specified layers

For example:

A network with 784 input-nodes, 2 hidden layers with 1000 nodes each and 10 output-nodes

> createNeuralNetwork [784,1000,1000,10]

Prediction

predict :: NeuralNetwork -> Matrix Float -> Matrix Float

predict takes an network and a fitting input and runs a forwardPass with these parameters. The resulting output Matrix is returned

Training functions

forwardPass :: NeuralNetwork -> Matrix Float -> [Matrix Float]

train :: NeuralNetwork -> [(Matrix Float, Matrix Float)] -> Float -> IO NeuralNetwork

train' :: NeuralNetwork -> [(Matrix Float, Matrix Float)] -> Float -> Float -> Int -> IO NeuralNetwork

data BackpropResult

update :: NeuralNetwork -> [(Matrix Float, Matrix Float)] -> Float -> NeuralNetwork

gradients :: [Matrix Float] -> [Matrix Float] -> [Matrix Float] -> Matrix Float -> [(Matrix Float, Matrix Float)]

Serializiation

serialize :: NeuralNetwork -> FilePath -> IO ()

serialize2 :: NeuralNetwork -> FilePath -> IO ()

deserialize :: FilePath -> IO NeuralNetwork

deserialize2 :: FilePath -> IO NeuralNetwork

Helper functions

sigmoid :: Float -> Float

sigmoid' :: Float -> Float

randomRMatrix :: Int -> Int -> (Float, Float) -> Matrix Float

zeroMatrix :: Int -> Int -> Matrix Float

multiplyElementwise :: Matrix Float -> Matrix Float -> Matrix Float

shuffleSamples :: [(Matrix Float, Matrix Float)] -> Int -> [(Matrix Float, Matrix Float)]

argmax :: Matrix Float -> Int

toCategorical :: Int -> Int -> [Float]